fix make file

+++ cpp01/ex05/Harl.cpp
@@ -6,7 +6,7 @@
 /*   By: lray <lray@student.42lausanne.ch >         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/31 14:06:38 by lray              #+#    #+#             */
-/*   Updated: 2024/01/31 15:07:29 by lray             ###   ########.fr       */
+/*   Updated: 2024/02/09 12:30:48 by lray             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,8 +26,9 @@ void Harl::complain(std::string level) const {
 		"WARNING",
 		"ERROR"
 	};
-	for (int i = 0; i < 4; i++) {
-		if (level == levels[i]) {
+	for (size_t i = 0; i < 4; i++) {
+		if (level == levels[i])
+		{
 			(this->*funcs[i])();
 			return;
 		}
@@ -39,20 +40,17 @@ void Harl::debug() const {
 		"7XL-double-cheese-triple-pickle-special-ketchup burger. " <<
 		"I really do!" << std::endl;
 }
-void Harl::info() const
-{
+void Harl::info() const {
 	std::cout << "I cannot believe adding extra bacon costs more money. " <<
 	"You didn't put enough bacon in my burger! " <<
 	"If you did, I wouldn't be asking for more!" << std::endl;
 }
-void Harl::warning() const
-{
+void Harl::warning() const {
 	std::cout << "I think I deserve to have some extra bacon for free. " <<
 	"I've been coming for years whereas you started working here " <<
 	"since last month." << std::endl;
 }
-void Harl::error() const
-{
+void Harl::error() const {
 	std::cout << "This is unacceptable! I want to speak to the manager now." <<
 	std::endl;
 }
